# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Video Topic Graph Platform API
  description: |
    Open-source platform for video topic graph generation.
    All endpoints require OIDC authentication via Bearer token.
  version: 1.0.0
  contact:
    name: Video Topic Graph Platform
    url: https://github.com/video-topic-graph/platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /search:
    post:
      summary: Semantic search across topics
      description: Search topics by semantic similarity to query
      tags: [Search]
      operationId: searchTopics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/deep:
    post:
      summary: Deep search with LLM reasoning
      description: |
        Performs deep search using LLM to reason across topics.
        May include topic synthesis and cross-reference analysis.
      tags: [Search]
      operationId: deepSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeepSearchRequest'
      responses:
        '200':
          description: Deep search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepSearchResponse'
...
  /jobs/{id}:
    get:
      summary: Get job status
      description: Check the status of a pipeline job
      tags: [Jobs]
      operationId: getJobStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{id}/cancel:
    post:
      summary: Cancel a running job
      tags: [Jobs]
      operationId: cancelJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job cannot be cancelled
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC Bearer token from Keycloak
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    PipelineConfig:
      type: object
      description: Override default pipeline configuration
      properties:
        asr_model:
          type: string
          enum: [whisper, faster-whisper, whisper-cpp]
        asr_language:
          type: string
          description: ISO 639-1 language code (auto-detect if omitted)
        enable_diarization:
          type: boolean
          default: false
        enable_scene_detection:
          type: boolean
          default: false
        topic_levels:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        embedding_model:
          type: string
          enum: [all-MiniLM-L6-v2, all-mpnet-base-v2, ollama-nomic]
        llm_model:
          type: string
          enum: [llama2, mistral, mixtral, ollama-phi]
        snippet_quality:
          type: string
          enum: [low, medium, high]
          default: medium
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 1000
        video_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Limit search to specific videos
        filters:
          type: object
          properties:
            level:
              type: integer
            min_importance:
              type: number
              minimum: 0
              maximum: 1
            date_from:
              type: string
              format: date-time
            date_to:
              type: string
              format: date-time
        options:
          type: object
          properties:
            include_transcripts:
              type: boolean
              default: false
            include_edges:
              type: boolean
              default: false
            top_k:
              type: integer
              default: 10
              maximum: 100
    DeepSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        video_ids:
          type: array
          items:
            type: string
            format: uuid
        context:
          type: object
          properties:
            include_synthesis:
              type: boolean
              default: true
            include_cross_references:
              type: boolean
              default: true
            max_topics_to_analyze:
              type: integer
              default: 20
...
    SearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
        took_ms:
          type: integer
    SearchResult:
      type: object
      properties:
        topic:
          $ref: '#/components/schemas/TopicNode'
        score:
          type: number
        matched_transcript:
          type: string
          nullable: true
        highlight:
          type: string
          nullable: true
    DeepSearchResponse:
      type: object
      properties:
        query:
          type: string
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        synthesis:
          type: string
          nullable: true
        cross_references:
          type: array
          items:
            type: object
            properties:
              topic_a_id:
                type: string
              topic_b_id:
                type: string
              relationship:
                type: string
        took_ms:
          type: integer
...
    JobStatusResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [video_analysis, export, snippet_generation]
        status:
          type: string
          enum: [pending, running, paused, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 100
        current_step:
          type: string
          nullable: true
        steps:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              status:
                type: string
              started_at:
                type: string
                format: date-time
                nullable: true
              completed_at:
                type: string
                format: date-time
                nullable: true
              error:
                type: string
                nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
        manifest:
          $ref: '#/components/schemas/ArtifactManifest'
          nullable: true
...
    TopicNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        level:
          type: integer
        start_ts:
          type: number
          description: Start timestamp in seconds
        end_ts:
          type: number
          description: End timestamp in seconds
        title:
          type: string
        summary:
          type: string
        keywords:
          type: array
          items:
            type: string
        parent_ids:
          type: array
          items:
            type: string
            format: uuid
        child_ids:
          type: array
          items:
            type: string
            format: uuid
        importance_score:
          type: number
          minimum: 0
          maximum: 1
        cluster_id:
          type: string
          nullable: true
...
    ArtifactManifest:
      type: object
      description: Pipeline artifact manifest passed between steps
      properties:
        video_id:
          type: string
          format: uuid
        graph_version_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        paths:
          type: object
          properties:
            video_original:
              type: string
            video_normalized:
              type: string
            audio_wav:
              type: string
            transcript:
              type: string
            topics:
              type: string
            embeddings:
              type: string
            graph:
              type: string
            snippets:
              type: array
              items:
                type: string
            exports:
              type: array
              items:
                type: string
        metrics:
          type: object
          properties:
            duration_s:
              type: number
            transcript_segments:
              type: integer
            topic_count:
              type: integer
            edge_count:
              type: integer
            processing_time_ms:
              type: number
        config_snapshot:
          $ref: '#/components/schemas/PipelineConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        request_id:
          type: string